Complete video 37

+++ backend/src/resolvers/Mutation.js
@@ -48,12 +48,17 @@ const Mutations = {
   },
 
   async deleteItem(parent, args, ctx, info) {
-    throw new Error('You are not allowed');
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
     // 2. check if they own that item, or have the permission
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN, ITEMDELETE'].includes(permission),
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permissions to do that");
+    }
     // 3. delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
